+++ src/main/java/com/outsider/masterofpredictionbackend/notification/query/controller/NotificationController.java
@@ -8,23 +8,30 @@ import com.outsider.masterofpredictionbackend.user.command.application.dto.Custo
 import com.outsider.masterofpredictionbackend.util.UserId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import org.springframework.data.web.PageableDefault;
+
 @RestController
 @RequestMapping("/api/v1/notifications")
 public class NotificationController {
 
     private final NotificationQueryService notificationService;
+
     public NotificationController(NotificationQueryService notificationService) {
         this.notificationService = notificationService;
     }
+
     @GetMapping
     public ResponseEntity<Page<NotificationResponseDTO>> getUserNotifications(
-            @UserId CustomUserInfoDTO userId, Pageable pageable) {
+            @UserId CustomUserInfoDTO userId,
+            @PageableDefault(size = 10, page = 0, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
+
         Page<NotificationResponseDTO> notifications = notificationService.getNotificationsByUserId(userId.getUserId(), pageable);
         if (notifications.isEmpty()) {
             return ResponseEntity.noContent().build();
