+++ src/main/java/com/outsider/masterofpredictionbackend/ranking/command/domain/aggregate/UserRanking.java
@@ -1,4 +1,4 @@
-package com.outsider.masterofpredictionbackend.user.query.tier.command;
+package com.outsider.masterofpredictionbackend.ranking.command.domain.aggregate;
 
 import jakarta.persistence.*;
